requirements.txt
google-api-python-client==2.118.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pyyaml==6.0.1
pytest==8.0.0
watchdog==3.0.0
